+++ src/main/java/temp/currency/CurrencyService.java
@@ -2,8 +2,9 @@ package temp.currency;
 
 import temp.currency.dto.Currency;
 import java.io.IOException;
+import java.math.BigDecimal;
 import java.util.HashMap;
 
 public interface CurrencyService {
-    HashMap<String, Double> getRate(Currency currency) throws IOException;
+    HashMap<String, BigDecimal> getRate(Currency currency) throws IOException;
 }
